Review feedback and a11y fixes
