+++ apps/src/templates/levelSummary/MultiResponses.jsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, {useMemo} from 'react';
 import PropTypes from 'prop-types';
 import styles from './summary.module.scss';
 import {Chart} from 'react-google-charts';
@@ -16,15 +16,6 @@ const MultiResponses = ({scriptData, showCorrectAnswer = false}) => {
       acc[letter] = acc[letter] ? acc[letter] + 1 : 1;
       return acc;
     }, {}),
-    /*
-    // FIXME: test data
-    ...{
-      A: 1,
-      B: 4,
-      C: 19,
-      D: 0,
-    },
-    */
   });
 
   const multiChartData = (data, highlightCorrect = false) => [
@@ -42,82 +33,78 @@ const MultiResponses = ({scriptData, showCorrectAnswer = false}) => {
   ];
 
   const answers = scriptData.level.properties.answers;
-  const answerData = multiAnswerCounts(scriptData.responses, answers.length);
-  const answerMax = Math.max(...Object.values(answerData));
-  const correctAnswers = answers.reduce((acc, cur, i) => {
-    if (cur.correct) {
-      acc.push([...LETTERS][i]);
-    }
-    return acc;
-  }, []);
-
-  // FIXME: Is this too weird?
-  const correctAnswerElement = document.getElementById(
-    'summary-correct-answer'
+  const answerData = useMemo(
+    () => multiAnswerCounts(scriptData.responses, answers.length),
+    [scriptData.responses, answers.length]
   );
-  if (showCorrectAnswer) {
-    correctAnswerElement.classList.add(styles.correctAnswersContainer);
-    correctAnswerElement.classList.remove('hide');
-  } else {
-    correctAnswerElement.classList.add('hide');
-    correctAnswerElement.classList.remove(styles.correctAnswersContainer);
-  }
+  const answerMax = useMemo(
+    () => Math.max(...Object.values(answerData)),
+    [answerData]
+  );
+  const correctAnswers = useMemo(
+    () =>
+      answers.reduce((acc, cur, i) => {
+        if (cur.correct) {
+          acc.push([...LETTERS][i]);
+        }
+        return acc;
+      }, []),
+    [answers]
+  );
 
   return (
     <div className={styles.multiChart}>
-      {scriptData.responses.length > 0 && (
-        <Chart
-          chartType="ColumnChart"
-          width="100%"
-          height="340px"
-          data={multiChartData(
-            answerData,
-            showCorrectAnswer ? correctAnswers : false
-          )}
-          options={{
-            bar: {
-              groupWidth: '80%',
+      <Chart
+        chartType="ColumnChart"
+        width="100%"
+        height="340px"
+        data={multiChartData(
+          answerData,
+          showCorrectAnswer ? correctAnswers : false
+        )}
+        options={{
+          bar: {
+            groupWidth: '80%',
+          },
+          colors: [color.brand_primary_default],
+          annotations: {
+            alwaysOutside: true,
+            textStyle: {
+              fontSize: 24,
+              color: 'black',
             },
-            colors: [color.brand_primary_default],
-            annotations: {
-              alwaysOutside: true,
-              textStyle: {
-                fontSize: 24,
-                color: 'black',
-              },
-              stem: {
-                color: 'none',
-              },
+            stem: {
+              color: 'none',
             },
-            legend: 'none',
-            vAxis: {
-              gridlines: {
-                count: 0,
-              },
-              ticks: [],
-              // Give some padding, so the annotation draws correctly
-              // on top of the highest bar.
-              maxValue: Math.ceil(answerMax * 1.2),
+          },
+          legend: 'none',
+          vAxis: {
+            gridlines: {
+              count: 0,
             },
-            hAxis: {
-              textStyle: {
-                fontSize: 24,
-              },
+            ticks: [],
+            // Give some padding, so the annotation draws correctly
+            // on top of the highest bar. Never set the max to 0.
+            maxValue: Math.ceil(answerMax * 1.2) || 1,
+          },
+          hAxis: {
+            textStyle: {
+              fontSize: 24,
             },
-            chartArea: {
-              width: '100%',
-              height: '90%',
-              left: 0,
-              right: 0,
-              top: 0,
-            },
-            tooltip: {
-              trigger: 'none',
-            },
-            enableInteractivity: false,
-          }}
-        />
-      )}
+          },
+          chartArea: {
+            width: '100%',
+            height: '90%',
+            left: 0,
+            right: 0,
+            top: 0,
+          },
+          tooltip: {
+            trigger: 'none',
+          },
+          enableInteractivity: false,
+        }}
+      />
     </div>
   );
 };
